#1449 Update commands.yml and make getter/setter in sync for new delay property

+++ src/main/java/fr/xephi/authme/settings/commandconfig/Command.java
@@ -61,7 +61,7 @@ public class Command {
         return delay;
     }
 
-    public void setDelay(int delay) {
+    public void setDelay(long delay) {
         this.delay = delay;
     }
